Fix type issue

+++ src/turret_node.cpp
@@ -60,7 +60,7 @@ enum class TurretStates
     PREPARE_CLIMB
 };
 
-static constexpr float SHOOTER_RPM_DELTA = 150;
+static constexpr double SHOOTER_RPM_DELTA = 150;
 static constexpr float HOOD_DEG_DELTA = 2;
 static constexpr float TURRET_YAW_DEG_DELTA = 2;
 static constexpr float SHOOTER_RPM_FILTER_TIME = 0.0;
